adapted tests to new repository structure
