+++ code/TS1_DivisionByZero/FunctionCall/java/DivideByZero.java
@@ -9,13 +9,6 @@
 
 public class DivideByZero {
 
-  public int divByZeroLocal(String s) {
-    int denominator = 0;
-    int nominator = 10;
-    int result = nominator / denominator;
-    return result;
-  }
-
   public int divideByZeroInterProc(int denominator) {
     return 10 / denominator;
   }
@@ -25,15 +18,4 @@ public class DivideByZero {
     return divideByZeroInterProc(0);
   }
 
-  // divide by zero with static fields
-  private static int x;
-
-  public void setXToZero() {
-    x = 0;
-  }
-
-  public int divideByZeroWithStaticField() {
-    setXToZero();
-    return divideByZeroInterProc(x);
-  }
 }
