--- code/TS2_RecursiveCall/README.md
@@ -1,3 +1,0 @@
-This directory contains tests and code samples used by Facebook Infer static analysis tool.
-
-https://github.com/facebook/infer
