+++ code/TS2_RecursiveCall/derived_again/c++/DerivedAgainExample.cpp
