+++ code/TS2_RecursiveCall/derived_again/java/README.md
@@ -1,0 +1,12 @@
+# A derived class inherits a pair of mutually recursive methods.
+
+```
+# Name        : A derived class DerivedAgain inherits a pair of mutually recursive methods from Derived.
+# FailureType : StackOverflowError
+# ErrorType   : Programming error
+# Source      : https://github.com/Sitiritis/eo-static-analyzer/blob/demo/sandbox/src/main/resources/mutual_recursion_non_term_derived_again.eo
+# CodeType    : Artificial
+# Lines       : (32, 39)
+```
+
+## *NOTE*: the groups of numbers in parentheses indicate the lines containing mutually recursive calls.
