+++ code/TS2_RecursiveCall/example/java/README.md
@@ -1,0 +1,13 @@
+# Mutual Recursion Example
+
+```
+# Name        : Mutual Recursion as a result of inheritance
+# FailureType : StackOverflowError
+# ErrorType   : Programming error
+# Source      : https://github.com/Sitiritis/eo-static-analyzer/tree/demo/sandbox/src/main/resources
+# CodeType    : Artificial
+# Lines       : (34, 41)
+```
+
+## *NOTE*: the groups of numbers in parentheses indicate the lines containing mutually recursive calls.
+
