+++ code/TS2_RecursiveCall/factory/c++/Factory.cpp
