+++ code/TS2_RecursiveCall/long_chain/c++/LongChain.cpp
