+++ code/TS2_RecursiveCall/two_derived/c++/TwoDerived.cpp
