+++ code/TS2_RecursiveCall/two_derived_explicit_decoration/c++/TwoDerivedExplicitDecoration.cpp
