+++ code/TS2_RecursiveCall/two_derived_explicit_decoration/java/README.md
@@ -1,0 +1,13 @@
+# 2 classes implement one base method in such a way that the other base method becomes mutually recursive with the first one.
+
+```
+# Name        : 2 classes implement one base method in such a way
+                that the other base method becomes mutually recursive with the first one.
+# FailureType : StackOverflowError
+# ErrorType   : Programming error
+# Source      : https://github.com/Sitiritis/eo-static-analyzer/blob/demo/sandbox/src/main/resources/mutual_rec_non_term_2_derived_explicit_decoration.eo
+# CodeType    : Artificial
+# Lines       : (34, 42), (34, 50)
+```
+
+## *NOTE*: the groups of numbers in parentheses indicate the lines containing mutually recursive calls.
