+++ code/TS3_NullPointerException/InferNullableTests/java/README.md
@@ -1,0 +1,14 @@
+# Native methods NPE tests
+
+```
+# Name        : Infer divide by zero tests
+# FailureType : Divide by zero
+# ErrorType   : Programming error
+# Source      : FB Infer
+# CodeType    : Artificial
+# Lines       : 15, 25, 37
+```
+
+Extracted from FB Infer tests located at:
+
+- https://github.com/facebook/infer/blob/master/infer/tests/codetoanalyze/java/biabduction/DivideByZero.java
