+++ code/TS3_NullPointerException/NullPointerExceptions/Native/java/InferNativeNPETests.java
