+++ code/TS3_NullPointerException/NullPointerExceptions/Nullable/java/README.md
@@ -1,0 +1,14 @@
+# Nullable tests
+
+```
+# Name        : Infer nullable fields tests
+# FailureType : Null Pointer Exceptions
+# ErrorType   : Programming error
+# Source      : FB Infer
+# CodeType    : Artificial
+# Lines       : 8, 12, 17, 21
+```
+
+Extracted from FB Infer tests located at:
+
+- https://github.com/facebook/infer/blob/master/infer/tests/codetoanalyze/java/biabduction/NullPointerExceptions.java
