+++ code/TS3_NullPointerException/NullPointerExceptions/Trivial/java/InferTrivialNPETests.java
