+++ code/TS3_NullPointerException/NullPointerExceptions/WithStandardLib/java/InferSTDLibNPETests.java
